Fixed the resizing bug for tenshi front end.
The default installation of Foundation will break when the screen
becomes too small. Foundation allows for more customization using what
is known as "Sass" or "SCSS". Unfortunately, to use Sass you can't use
the default installation of Foundation. Instead you must follow these
instructions:
http://foundation.zurb.com/docs/sass.html
To incorporate this into the existing project, a new "Foundation"
project was created following the links instructions, and the project
was "merged" into the existing project. Finally, the line:
// $topbar-breakpoint: emCalc(940px);
was uncommented and emCalc(940px) was changed to 0px.
This fixes the resizing issue.

Change-Id: I417e2a4359c2777cfd0a12c4f8e7ad88a4caf85a