Very primitive implementation of block snapping
Blocks can snap to blocks that are within a certain radius of their dropped
location. The block being dragged and the block being snapped to need to be
at least half-overlapped in the horizontal direction in order to snap.
If multiple are found, it simply stupidly picks the first one found by
iterating the block list.

Change-Id: Ief6bda2b4f7422a008c8fbc02ab16b078d426199
Reviewed-on: https://robertou.com/gerrit/114
Tested-by: Jenkins CI
Reviewed-by: Kyle Zentner <kylezentner@pioneers.berkeley.edu>