Review fixes
I believe it should all be fixed now. All the capacitors have values and
all name/value positions are standardized (with the exception of COUT2
which hadn't sufficient room around the sides to place text.) Grizzly's
solution, while elegant, I don't think works as well here because we
don't have a bunch of components with the same value next to each other
as Grizzly does. Is there a way I can still do that here and clean up
the silkscreen?

In terms of the holes, I simply opted to move the hole in question to
the middle of the board to obviate the problem of the switching
regulator overlap altogether. Also, the holes at the bottom are at the
same height. Will enclosures people kill me for putting the holes at
.235 inches from the bottom of the board? I think the top-right mounting
hole actually misses the uMaestro altogether (the board was built to
accommodate a larger Maestro) and can go straight through into the
board. I hope. We'll have to test it.