LED system update
Howdy review team. I have cobbled together this possible fix for the
serial indicator LED problem we had discussed, pursuant to information
found on the following forum thread:
http://forum.arduino.cc/index.php/topic,12216.0.html
Let me know if this even works. I know I need to add a chamfer to the
LED/RX line branch, but I'm more interested in the schematic aspect of
this for the time being. Also, in terms of what we want to do for a test
jig, might it be helpful to build some auxiliary test points into the
board? If so, where?