Inital routing of controller board
Non-optimized. Currently debugging massive amount of clearance errors.

Justifications for approved DRC errors:
* Dimension errors: The Anderson mechanical strength pads are not sensitive to
issues such as being accidentally cut into. The USB connector is required by
the datasheet and is unlikely to have issues (the DRC is very conservative).
The SD connector is not required but should be OK because it is actually
further in than the USB connector
* Overlap errors: The holes in the V-reg are for thermal purposes. The
holes/pads on the Anderson are for mechanical strength.

Change-Id: I675fb15063f77c34d2b80af91de120a98d90e861
Reviewed-on: https://robertou.com/gerrit/201
Tested-by: Jenkins CI
Reviewed-by: Kyle Zentner <kylezentner@pioneers.berkeley.edu>