Add a require-like system for sharing data between iframes
Each frame defines an exports object in its global scope, which can then
be loaded by other frames.

The system is intended so the individual HTML pages in the frames can be
loaded standalone in a web browser, with a graceful fallback to the
subset of functionality that doesn't require interaction with other
frames.

Change-Id: I19427ff8f7517a2d88bd041ea3b0c77ab58f6ee9
Reviewed-on: https://robertou.com/gerrit/244
Tested-by: Jenkins CI
Reviewed-by: Robert Ou <robertou@pioneers.berkeley.edu>