Add back/forward/reload buttons to debug footer
In order for history to properly work, the type of the main browser needs
to be changed to a content browser instead of a chrome browser. It is not
clear if this has a workaround or not. However, it appears that a content
browser restricts access to the outer chrome, but it does not revoke
chrome-level privileges.
TODO: Investigate what gives/revokes chrome privileges.

Change-Id: Ice75c62a72306f173ac44fbb7d119d9d684dd462
Reviewed-on: https://robertou.com/gerrit/304
Tested-by: Jenkins CI
Reviewed-by: Kyle Zentner <kylezentner@pioneers.berkeley.edu>