Refactor serial port code to use web workers.
The serial port code now works with Chrome Workers, which is basically
idential to web workers except that it also has access to js-ctypes.
Unfortunately, this code breaks support with being required via the
normal Jetpack require system. However, this should be an acceptable
loss because we will almost certainly eventually run this code in a
worker so that it can run in the background while other parts of the
app are running.

Change-Id: I453bd0e47d3900b47648948590a1fcb61da1f1cc