Initial commit of PiELES
PiELES (Pioneers in Engineering Lightweight Extensible System) is the
successor to PiEMOS for driving robots. Unlike PiEMOS, it is not
designed as a stand-alone application, but is instead a "widget" that
lives within the Tenshi IDE. It will eventually have a page to
configure it and icons that indicate its status, but it will not have
any display. The feedback graphs and other visualizations will take the
place of what used to be displayed by PiEMOS.

PiELES sends updates to the robot at a constant update rate rather than
an inconsistent rate proportional to controller updates. This is intended
to make control theory analysis of the system easier. It is expected that
this will be fine once the radio is upgraded.

This initial version of PiELES supports mapping controller inputs to the
legacy packet format. It does not yet support controlling the simulator.
All of these features will be added at a later time. At this time, PiELES
is capable of driving Tenshi Test Bot.

Change-Id: I56450db2dd46647ee5df3d3e3362641e64b4098a