Added Mac OS X support for serial ports
This was tested on Mac OS X 10.9 x86_64 to at least successfully send
data out a serial port. On Mac OS X, it seems you should use the
/dev/cu.* devices rather than the /dev/tty.* devices. This was not tested
on Mac OS X i386, but it should work. It is not guaranteed to work on
Mac OS X PPC, which is even more unsupported by our project than Mac OS x86.

Change-Id: I09388d0fbee079d4ff71a5f00d9224c58df3a90d