Add the Open Sans font family
The entire family is included in the TrueType format. No effort is made
to optimize space or loading times, so this does not use typical web
techniques such as compressed fonts (woff format), selective importing,
or including only a subset of the glyphs.

This change also introduces a common.css file that will be imported from
throughout the application, and should be used for application-wide
style settings.