Reduce the number of bits in the "sample" field
Previously, the byte range for active mode packet second bytes
could overlap with the range used for maintainance types.
By reducing the range of the "sample" bitfield, this overlap is
avoided. One problem with this is that it makes "One subchunk ever
n samples" even tricker than before, but this a) can be worked
around and b) will not be supported initially.

Change-Id: Idcdc04e797f4b6ee2c78601229675b1ab36e8847