Fix EAGLE race condition and refactored scripts
EAGLE has a race condition where it tries to rewrite its config file every
time it quits. It does this by writing to .~~~ and then renaming that to
.eaglerc. However, if multiple processes quit at the same time, some of
them can fail. This can cause EAGLE to hang with an error message in Xvfb.
This fixes the problem by simply preventing the rename.

This commit also refactors the EAGLE-related python scripts to not duplicate
nearly as much code. Also, there is no longer a stupid wait for Xvfb to start
(now it checks for the X socket to be created). The code also fixes killing
Xvfb and leaving lock files all over /tmp.