Set up automated testing in XULRunner
This is a basic test launcher (no true test framework) that loads all of the
tests in the test/tests directory as CommonJS modules and calls their
exported run functions. It is assumed that the run function will return a
true/false value. The test log will be printed out and the Python wrapper
script will exit with a 0 status only if the tests all pass.

Also silence useless Xvfb output.