Add the ability to run HTML tests
These tests load an HTML page into a chrome-privileged <browser>. The test
can then do whatever it chooses to. It is most likely that the test will
load Jasmine and run further tests. These tests obviously have access to
the DOM. In order to indicate success/failure of one of these tests to the
test framework, raise a TenshiTest custom event with the detail field set
to the success/failure status.