Switch to persistent divs for main ui
The angular-based framework will preload all of the elements into divs
and toggle their layout and visibility to achieve module switching. The
broken layout of this version will be fixed once angular has been added.

Disabling resource-intensive features (such as the simulator's physics
engine and rendering) when not visible will be handled within the
individual modules.

The robot maker is temporarily disabled because the simulator and the
maker use the same variable to store the window object.