Added sensor and actuator support to the mailbox API
Sensors and actuators are treated almost identically to regular mailboxes.
They are stored in a global table and are identified by an ID. The ID can
be e.g. directly a smartsensor ID number or a special string identifying
PiELES data. Although sensors are not mailboxes in the student-facing API,
they are mailboxes that get a value every sample in the underlying
implementation. This fact will be hidden by the student-facing get_device
function.