Change FreeRTOS allocator to heap_1.c
This is the naive memory allocator that allocates out of an array and does not
support freeing. However, this allocator is now only used for allocating
FreeRTOS objects like TCBs and stacks. None of these ever get freed.

Note that now it is not possible to DMA into buffers on the stack.