Run code from ROM patch 1/5 -- ZIO "direct mode"
This patch series is derived from some of the work done on eLua. Specifically,
this set of patches is derived from
https://github.com/elua/elua/commit/d54659b5723bcd2b1e3900362398c72c18a9aa0b.

This commit changes lzio to handle a "direct mode" where it does not actually
buffer and read data but instead directly exposes the backing pointer (which
is most likely pointing into ROM). Note that this patch does not cover all
use cases of lzio (specifically what is needed for the lexer/parser) and only
can handle the use cases in lundump.