Implement eLua pseudo-RO strings
These strings contain the actual string bytes in ROM. However, the TString
object is still stored in RAM and contains a pointer to the ROM data where
the string bytes would normally live. Note that if a pseudo-RO string gets
GC'd and recreated, it most likely will no longer be a pseudo-RO string.