New Arduino shield PCB that allows an Arduino to be a smart device
This board allows the Hibike project to develop code for the different smart
devices without requiring an unpowered smart device.  It connects between
the Arduino and either unpowered sensors, the powered smart device board,
the battery buzzer, or the team flag through male/female header pins or pogo
pins.

This can also potentially be used as a test jig for programming the semart
device through the ISP connector just by reprogramming the Arduino.